% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{activity}
\alias{activity}
\title{Computes activity of a Protein-protein interaction network associated with a biological process for each cell}
\usage{
activity(expression_matrix, adjacency_edges, normalization = FALSE)
}
\arguments{
\item{expression_matrix}{Normalized expression matrix with cells as columns and genes as rows (matrix or data frame)}

\item{adjacency_edges}{Adjacency edges of the protein protein interaction network. Data frame of two columns (nodes) and N rows (edges). Can be custom-made or any PPIN included in the package}

\item{normalization}{If TRUE, normalize by the network size (number of edges). Default is FALSE}
}
\value{
A vector with the PPIN activity values for each cell
}
\description{
Computes activity of a Protein-protein interaction network associated with a biological process for each cell
}
