% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acute_inflammatory_response_edges}
\alias{acute_inflammatory_response_edges}
\title{PPIN associated with acute inflammatory response BP}
\format{
A data frame with 208 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(acute_inflammatory_response_edges) # Lazy loading
}
\usage{
acute_inflammatory_response_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with acute inflammatory response BP (GO:0002526), each row represents an interaction
}
\keyword{datasets}
