% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{differentiation_edges}
\alias{differentiation_edges}
\title{PPIN associated with cell differentiation BP}
\format{
A data frame with 74448 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(differentiation_edges) # Lazy loading
}
\usage{
differentiation_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with cell differentiation (GO:0030154), each row represents an interaction
}
\keyword{datasets}
