% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dna_replication_edges}
\alias{dna_replication_edges}
\title{PPIN associated with DNA replication BP}
\format{
A data frame with 2010 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(dna_replication_edges) # Lazy loading
}
\usage{
dna_replication_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with DNA replication (GO:0006260), each row represents an interaction
}
\keyword{datasets}
