% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{full_human_ppin}
\alias{full_human_ppin}
\title{Full human protein-protein interaction network}
\format{
Data frame with 2080394 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
\url{https://www.pathwaycommons.org/}
data(full_human_ppi) # Lazy loading
}
\usage{
full_human_ppin
}
\description{
A data frame containing an adjacency list of all the human protein-protein interactions, each row represents an interaction
}
\keyword{datasets}
