% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hormone_secretion_edges}
\alias{hormone_secretion_edges}
\title{PPIN associated with hormone secretion BP}
\format{
A data frame with 184 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(hormone_secretion_edges) # Lazy loading
}
\usage{
hormone_secretion_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with hormone secretion  (GO:0046879), each row represents an interaction
}
\keyword{datasets}
