% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{migration_edges}
\alias{migration_edges}
\title{PPIN associated with cell migration BP}
\format{
A data frame with 11234 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(migration_edges) # Lazy loading
}
\usage{
migration_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with cell migration (GO:0016477), each row represents an interaction
}
\keyword{datasets}
