% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mitotic_cell_cycle_edges}
\alias{mitotic_cell_cycle_edges}
\title{PPIN associated with mitotic cell cycle process BP}
\format{
A data frame with 4037 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(mitotic_cell_cycle_edges) # Lazy loading
}
\usage{
mitotic_cell_cycle_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with mitotic cell cycle process (GO:1903047), each row represents an interaction
}
\keyword{datasets}
