% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proliferation_edges}
\alias{proliferation_edges}
\title{PPIN associated with cell population proliferation BP}
\format{
A data frame with 6869 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(proliferation_edges) # Lazy loading
}
\usage{
proliferation_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with cell population proliferation (GO:0008283), each row represents an interaction
}
\keyword{datasets}
