% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stem_differentiation_edges}
\alias{stem_differentiation_edges}
\title{PPIN associated with stem cell differentiation  BP}
\format{
A data frame with 781 rows (edges) and 2 columns (nodes):
\describe{
\item{node1}{1st edge node}
\item{node2}{2nd edge node}
}
}
\source{
<https://www.ebi.ac.uk/QuickGO/ and https://www.pathwaycommons.org/>
data(stem_differentiation_edges) # Lazy loading
}
\usage{
stem_differentiation_edges
}
\description{
Data frame containing an adjacency list of the Protein-protein interaction network (PPIN) associated with stem cell differentiation (GO:0048863), each row represents an interaction
}
\keyword{datasets}
